+++ src/components/settings.tsx
@@ -2,16 +2,19 @@ import { Description } from '@radix-ui/react-dialog';
 import { useEffect, useId, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { updaterConfig } from '@/config/updater';
+import type { UpdateCheckResult } from '@/lib/update-checker';
+import { checkForUpdate, normalizeVersion } from '@/lib/update-checker';
 import {
     getActiveProjectFromState,
     useCaaSConfigStore,
 } from '@/stores/caas-config-store';
 import { useSettingsStore } from '@/stores/settings-store';
-import { type ModeType, modeOptionTexts, modeTypes } from '@/types/mode';
-import { type StageType, stageOptionTexts } from '@/types/stage';
+import type { ModeType } from '@/types/mode';
+import type { StageType } from '@/types/stage';
 import { version } from '../../package.json';
 import Icon from './icons/icon';
 import ManageProjectsDialog from './manage-projects-dialog';
+import { ProjectSettingsSection } from './project-settings-section';
 import { Button } from './ui/button';
 import { CheckboxFieldWithTooltip } from './ui/checkbox-field-with-tooltip';
 import {
@@ -22,50 +25,6 @@ import {
     DialogTitle,
     DialogTrigger,
 } from './ui/dialog';
-import {
-    Select,
-    SelectContent,
-    SelectItem,
-    SelectTrigger,
-    SelectValue,
-} from './ui/select';
-
-type UpdateCheckResult =
-    | {
-          status: 'upToDate';
-          latestVersion: string;
-      }
-    | {
-          status: 'updateAvailable';
-          latestVersion: string;
-          releaseUrl: string;
-      }
-    | {
-          status: 'error';
-      };
-
-const normalizeVersion = (rawVersion: string): string =>
-    rawVersion.trim().replace(/^v/i, '').split('-')[0].split('+')[0];
-
-const compareVersions = (leftVersion: string, rightVersion: string): number => {
-    const leftParts = leftVersion.split('.').map((part) => Number(part));
-    const rightParts = rightVersion.split('.').map((part) => Number(part));
-    const maxLength = Math.max(leftParts.length, rightParts.length);
-
-    for (let index = 0; index < maxLength; index += 1) {
-        const leftPart = Number.isFinite(leftParts[index])
-            ? leftParts[index]
-            : 0;
-        const rightPart = Number.isFinite(rightParts[index])
-            ? rightParts[index]
-            : 0;
-
-        if (leftPart > rightPart) return 1;
-        if (leftPart < rightPart) return -1;
-    }
-
-    return 0;
-};
 
 function Settings() {
     const { t } = useTranslation();
@@ -124,6 +83,7 @@ function Settings() {
     const [localCount, setLocalCount] = useState(count);
 
     // Sync local state from store when dialog opens
+    // biome-ignore lint/correctness/useExhaustiveDependencies: We just need to set the active selection when the settings are open
     useEffect(() => {
         if (isSettingsOpen) {
             setLocalCustomerName(activeSelection?.customerName ?? '');
@@ -135,7 +95,6 @@ function Settings() {
             setLocalRep(rep);
             setLocalCount(count);
         }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [isSettingsOpen]);
 
     // Derive options from local state
@@ -168,8 +127,6 @@ function Settings() {
 
     const handleSave = (e: React.FormEvent<HTMLFormElement>) => {
         e.preventDefault();
-        // Call all three unconditionally in order — each reads from the store state
-        // left by the previous call, so they must chain: customer → stage → project
         setActiveCustomer(localCustomerName);
         setActiveStage(localStage);
         setActiveProject(localProjectName);
@@ -195,43 +152,16 @@ function Settings() {
         setIsManageProjectsOpen(true);
     };
 
-    const checkForUpdate = async () => {
+    const handleCheckForUpdate = async () => {
         setIsCheckingUpdate(true);
         setUpdateCheckResult(null);
-
         try {
-            const response = await fetch(updaterConfig.latestReleaseApiUrl);
-            if (!response.ok) {
-                throw new Error(
-                    `Update check failed with status ${response.status}`
-                );
-            }
-
-            const release = (await response.json()) as {
-                tag_name?: string;
-                html_url?: string;
-            };
-            const latestVersion = normalizeVersion(release.tag_name ?? '');
-            const currentVersion = normalizeVersion(version);
-
-            if (!latestVersion || !currentVersion) {
-                throw new Error('Invalid version payload');
-            }
-
-            const comparison = compareVersions(currentVersion, latestVersion);
-            if (comparison < 0) {
-                setUpdateCheckResult({
-                    status: 'updateAvailable',
-                    latestVersion,
-                    releaseUrl:
-                        release.html_url ?? updaterConfig.releasesPageUrl,
-                });
-            } else {
-                setUpdateCheckResult({
-                    status: 'upToDate',
-                    latestVersion,
-                });
-            }
+            const result = await checkForUpdate(
+                version,
+                updaterConfig.latestReleaseApiUrl,
+                updaterConfig.releasesPageUrl
+            );
+            setUpdateCheckResult(result);
         } catch {
             setUpdateCheckResult({ status: 'error' });
         } finally {
@@ -258,177 +188,23 @@ function Settings() {
                             {t('app.settings.dialog.subtitle')}
                         </Description>
 
-                        {/* Project section */}
-                        <div className="flex flex-col gap-2 mt-4">
-                            <div className="flex items-center justify-between">
-                                <h2 className="font-semibold">
-                                    {t('app.form.project')}
-                                </h2>
-                                <Button
-                                    type="button"
-                                    onClick={openManageProjectsDialog}
-                                    variant="ghost"
-                                    size="sm"
-                                    className="text-muted-foreground gap-1.5"
-                                >
-                                    <Icon
-                                        icon="plus-circle"
-                                        className="size-4"
-                                    />
-                                    {t('app.settings.dialog.manageProjects')}
-                                </Button>
-                            </div>
-
-                            <div className="grid grid-cols-3 gap-3">
-                                <div className="flex flex-col gap-1.5">
-                                    <span className="text-sm">
-                                        {t('app.form.customer')}
-                                    </span>
-                                    <Select
-                                        value={localCustomerName}
-                                        onValueChange={handleCustomerChange}
-                                    >
-                                        <SelectTrigger className="w-full">
-                                            <SelectValue
-                                                placeholder={t(
-                                                    'app.form.customer'
-                                                )}
-                                            />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            {customerOptions.map(
-                                                (customerName) => (
-                                                    <SelectItem
-                                                        key={customerName}
-                                                        value={customerName}
-                                                    >
-                                                        {customerName}
-                                                    </SelectItem>
-                                                )
-                                            )}
-                                        </SelectContent>
-                                    </Select>
-                                </div>
-                                <div className="flex flex-col gap-1.5">
-                                    <span className="text-sm">
-                                        {t('app.form.stage')}
-                                    </span>
-                                    <Select
-                                        value={localStage}
-                                        onValueChange={(value) =>
-                                            handleStageChange(
-                                                value as StageType
-                                            )
-                                        }
-                                        disabled={!localStageOptions.length}
-                                    >
-                                        <SelectTrigger className="w-full">
-                                            <SelectValue
-                                                placeholder={t(
-                                                    'app.form.stage'
-                                                )}
-                                            />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            {localStageOptions.map((stage) => (
-                                                <SelectItem
-                                                    key={stage}
-                                                    value={stage}
-                                                >
-                                                    {stageOptionTexts[stage]}
-                                                </SelectItem>
-                                            ))}
-                                        </SelectContent>
-                                    </Select>
-                                </div>
-                                <div className="flex flex-col gap-1.5">
-                                    <span className="text-sm">
-                                        {t('app.form.project')}
-                                    </span>
-                                    <Select
-                                        value={localProjectName}
-                                        onValueChange={setLocalProjectName}
-                                        disabled={!localProjectOptions.length}
-                                    >
-                                        <SelectTrigger className="w-full">
-                                            <SelectValue
-                                                placeholder={t(
-                                                    'app.form.project'
-                                                )}
-                                            />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            {localProjectOptions.map(
-                                                (project) => (
-                                                    <SelectItem
-                                                        key={
-                                                            project.projectName
-                                                        }
-                                                        value={
-                                                            project.projectName
-                                                        }
-                                                    >
-                                                        {project.projectName}
-                                                    </SelectItem>
-                                                )
-                                            )}
-                                        </SelectContent>
-                                    </Select>
-                                </div>
-                            </div>
-
-                            <div className="grid grid-cols-2 gap-3">
-                                <div className="flex flex-col gap-1.5">
-                                    <span className="text-sm">
-                                        {t('app.settings.dialog.locale.label')}
-                                    </span>
-                                    <Select
-                                        value={localLocale}
-                                        onValueChange={setLocalLocale}
-                                        disabled={!localProjectLocales.length}
-                                    >
-                                        <SelectTrigger className="w-full">
-                                            <SelectValue placeholder="locale" />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            {localProjectLocales.map((loc) => (
-                                                <SelectItem
-                                                    key={loc}
-                                                    value={loc}
-                                                >
-                                                    {loc}
-                                                </SelectItem>
-                                            ))}
-                                        </SelectContent>
-                                    </Select>
-                                </div>
-                                <div className="flex flex-col gap-1.5">
-                                    <span className="text-sm">
-                                        {t('app.settings.dialog.mode.label')}
-                                    </span>
-                                    <Select
-                                        value={localMode}
-                                        onValueChange={(value) =>
-                                            setLocalMode(value as ModeType)
-                                        }
-                                    >
-                                        <SelectTrigger className="w-full">
-                                            <SelectValue placeholder="mode" />
-                                        </SelectTrigger>
-                                        <SelectContent>
-                                            {modeTypes.map((modeType) => (
-                                                <SelectItem
-                                                    key={modeType}
-                                                    value={modeType}
-                                                >
-                                                    {modeOptionTexts[modeType]}
-                                                </SelectItem>
-                                            ))}
-                                        </SelectContent>
-                                    </Select>
-                                </div>
-                            </div>
-                        </div>
+                        <ProjectSettingsSection
+                            localCustomerName={localCustomerName}
+                            localStage={localStage}
+                            localProjectName={localProjectName}
+                            localLocale={localLocale}
+                            localMode={localMode}
+                            customerOptions={customerOptions}
+                            stageOptions={localStageOptions}
+                            projectOptions={localProjectOptions}
+                            localeOptions={localProjectLocales}
+                            onCustomerChange={handleCustomerChange}
+                            onStageChange={handleStageChange}
+                            onProjectChange={setLocalProjectName}
+                            onLocaleChange={setLocalLocale}
+                            onModeChange={setLocalMode}
+                            onManageProjects={openManageProjectsDialog}
+                        />
 
                         <hr className="border-t my-4" />
 
@@ -440,30 +216,42 @@ function Settings() {
                             <div className="flex items-center gap-8">
                                 <CheckboxFieldWithTooltip
                                     id={npId}
-                                    label={t('app.settings.dialog.queryParams.np.label')}
+                                    label={t(
+                                        'app.settings.dialog.queryParams.np.label'
+                                    )}
                                     checked={localNp}
                                     onCheckedChange={setLocalNp}
                                     tooltipIcon="information-circle"
                                     tooltipIconColor="text-blue-500"
-                                    tooltipText={t('app.settings.dialog.queryParams.np.tooltip')}
+                                    tooltipText={t(
+                                        'app.settings.dialog.queryParams.np.tooltip'
+                                    )}
                                 />
                                 <CheckboxFieldWithTooltip
                                     id={repId}
-                                    label={t('app.settings.dialog.queryParams.repPj.label')}
+                                    label={t(
+                                        'app.settings.dialog.queryParams.repPj.label'
+                                    )}
                                     checked={localRep}
                                     onCheckedChange={setLocalRep}
                                     tooltipIcon="exclamation-triangle"
                                     tooltipIconColor="text-orange-500"
-                                    tooltipText={t('app.settings.dialog.queryParams.repPj.tooltip')}
+                                    tooltipText={t(
+                                        'app.settings.dialog.queryParams.repPj.tooltip'
+                                    )}
                                 />
                                 <CheckboxFieldWithTooltip
                                     id={countId}
-                                    label={t('app.settings.dialog.queryParams.count.label')}
+                                    label={t(
+                                        'app.settings.dialog.queryParams.count.label'
+                                    )}
                                     checked={localCount}
                                     onCheckedChange={setLocalCount}
                                     tooltipIcon="information-circle"
                                     tooltipIconColor="text-blue-500"
-                                    tooltipText={t('app.settings.dialog.queryParams.count.tooltip')}
+                                    tooltipText={t(
+                                        'app.settings.dialog.queryParams.count.tooltip'
+                                    )}
                                 />
                             </div>
                         </div>
@@ -487,7 +275,7 @@ function Settings() {
                                 <Button
                                     type="button"
                                     variant="outline"
-                                    onClick={checkForUpdate}
+                                    onClick={handleCheckForUpdate}
                                     disabled={isCheckingUpdate}
                                 >
                                     <Icon
